[mojo-lang] Refactor the generation and verification of main (#16655)
`main` is currently in a weird spot for a few reasons:
* We implicitly import main only if the signature is correct, which
  can lead to confusing errors if the user doesn't know the expected
  signature.
* The user can define both `def main` and `fn main`, making it difficult
  for tools that need to call into main (e.g. the jit), as it needs to
  know about multiple potential ABIS.

This PR cleans this up by actually verifying the expected structure of
main, and by auto-generating a `main` wrapper depending on what the
user used to define their main. This shim also handles setting up argv,
and prints any exceptions in the case where main raises and an
exception isn't caught.

In addition to making the JIT's life easier, this will also greatly
simplify generating mojo executables.

MODULAR_ORIG_COMMIT_REV_ID: 66d99c39a22d9886c04b856e32803ac065dc42ad

+++ KGEN/lib/CompilerRT/System.cpp
@@ -71,12 +71,12 @@ KGEN_CompilerRT_GetArgV(ArgVList::RawList *result) {
 }
 
 COMPILERRT_EXPORT
-COMPILERRT_VISIBILITY_EXPORT void KGEN_CompilerRT_SetArgV(size_t argc,
+COMPILERRT_VISIBILITY_EXPORT void KGEN_CompilerRT_SetArgV(int argc,
                                                           char **argv) {
   ArgVList &argVList = ArgVList::get();
   argVList.args.resize(argc);
   argVList.argStrings.resize(argc);
-  for (size_t i = 0; i < argc; ++i)
+  for (int i = 0; i < argc; ++i)
     argVList.args[i] = argVList.argStrings[i] = argv[i];
 }
